fix: align read file error envelopes
Applies read-file recovery details through the shared CLI and MCP error paths so parity tests see identical FILE_NOT_FOUND envelopes.
